Cleanup and reorganize tests (#1026)
* Sort test files to matlab/tests/ and python/tests

* Ignore DeprecationWarnings

* Add model module import helper

* Switch to pytest and fixtures

* make tests independent
* avoid repeated import of same model
* make asserts less awkward

* Make testCoverage.py less terrible (Closes #1024)

* README

* Remove temporary files generated during tests (Closes #920)

* lint

* Update mtoc++ exclude list

* Replace tests/testCMakeCompilation.sh by pytest case
