+++ .coveragerc
@@ -1,0 +1,18 @@
+[run]
+
+source = amici
+omit =
+    */amici.py
+    */amici_without_hdf5.py
+    */ThirdParty/*
+    *.template.*
+parallel = true
+
+[report]
+
+exclude_lines =
+    pragma: no cover
+    raise
+    except:
+    import
+
