+++ .github/workflows/test-petab-test-suite.yml
@@ -47,18 +47,19 @@ jobs:
       run: |
         AMICI_PARALLEL_COMPILE=2 check_time.sh \
             install_sdist pip3 install -v --user \
-            $(ls -t python/sdist/dist/amici-*.tar.gz | head -1)[petab]
+            $(ls -t python/sdist/dist/amici-*.tar.gz | head -1)[petab,pysb]
 
     - name: Run PEtab-related unit tests
       run: |
-        pytest --cov-report=xml --cov=./ tests/testMisc.py
+        pytest --cov-report=xml --cov=./ python/tests/test_*petab*.py
 
     # retrieve test models
     - name: Download and run petab test suite
       run: |
         git clone --depth 1 https://github.com/petab-dev/petab_test_suite \
           && cd petab_test_suite; pip3 install -e .; cd .. \
-          && AMICI_PARALLEL_COMPILE=2 pytest -v --cov-report=xml --cov-append --cov=./ tests/petab_test_suite/
+          && AMICI_PARALLEL_COMPILE=2 pytest -v \
+            --cov-report=xml --cov-append --cov=amici tests/petab_test_suite/
 
     - name: Codecov
       uses: codecov/codecov-action@v1.0.6
