+++ .gitignore
@@ -104,29 +104,29 @@ python/testSpeedy.py
 python/test.txt
 
 tests/test/*
-tests/test_model_steadystate_scaled/*
-tests/test_model_presimulation/*
-tests/test_model_presimulation_pysb/*
-tests/test_model_presimulation_sbml/*
-tests/test_likelihoods/*
-tests/bax_pore_sequential_amici/*
-tests/bax_pore_amici/*
-tests/bngwiki_egfr_simple_amici/*
-tests/bngwiki_egfr_simple_deletemolecules_amici/*
-tests/earm_1_3_amici/*
-tests/earm_1_0_amici/*
-tests/expression_observables_amici/*
-tests/robertson_amici/*
-tests/tyson_oscillator_amici/*
-tests/bngwiki_simple_amici/*
-tests/bngwiki_enzymatic_cycle_mm_amici/*
-tests/move_connected_amici/*
-tests/michment_amici/*
-tests/kinase_cascade_amici/*
-tests/hello_pysb_amici/*
-tests/fricker_2010_apoptosis_amici/*
-tests/explicit_amici/*
-tests/fixed_initial_amici/*
+*/tests/test_model_steadystate_scaled/*
+*/tests/test_model_presimulation/*
+*/tests/test_model_presimulation_pysb/*
+*/tests/test_model_presimulation_sbml/*
+*/tests/test_likelihoods/*
+*/tests/bax_pore_sequential_amici/*
+*/tests/bax_pore_amici/*
+*/tests/bngwiki_egfr_simple_amici/*
+*/tests/bngwiki_egfr_simple_deletemolecules_amici/*
+*/tests/earm_1_3_amici/*
+*/tests/earm_1_0_amici/*
+*/tests/expression_observables_amici/*
+*/tests/robertson_amici/*
+*/tests/tyson_oscillator_amici/*
+*/tests/bngwiki_simple_amici/*
+*/tests/bngwiki_enzymatic_cycle_mm_amici/*
+*/tests/move_connected_amici/*
+*/tests/michment_amici/*
+*/tests/kinase_cascade_amici/*
+*/tests/hello_pysb_amici/*
+*/tests/fricker_2010_apoptosis_amici/*
+*/tests/explicit_amici/*
+*/tests/fixed_initial_amici/*
 tests/cpputest/writeResults.h5
 tests/cpputest/writeResults.h5.bak
 tests/sbml-test-suite/*
