+++ pytest.ini
@@ -1,4 +1,13 @@
 [pytest]
-# hundreds of SBML <=5.17 warnings
+
+addopts = -vv
+
 filterwarnings =
+    # hundreds of SBML <=5.17 warnings
     ignore:.*inspect.getargspec\(\) is deprecated.*:DeprecationWarning
+    # pysb warnings
+    ignore:the imp module is deprecated in favour of importlib.*:DeprecationWarning:pysb\.core
+    ignore:Using or importing the ABCs from 'collections' instead of from 'collections.abc' is deprecated since Python 3.3,and in 3.9 it will stop working.*:DeprecationWarning:pysb\.core
+    ignore:Model.initial_conditions will be removed in a future version. Instead, you can get a list of Initial objects with Model.initials.:DeprecationWarning:pysb\.core
+
+norecursedirs = .git amici_models build doc documentation matlab models ThirdParty amici sdist examples
