+++ python/sdist/setup.py
@@ -193,7 +193,8 @@ def main():
         setup_requires=['setuptools>=40.6.3'],
         python_requires='>=3.6',
         extras_require={
-            'petab': ['petab==0.1.5']
+            'petab': ['petab==0.1.5'],
+            'pysb': ['pysb>=1.11.0']
         },
         package_data={
             'amici': ['amici/include/amici/*',
