+++ python/tests/README.md
@@ -1,0 +1,9 @@
+# AMICI Python test suite
+
+This directory contains the AMICI Python test suite. For each python file
+of the AMICI package, there is a corresponding `test_${filename}.py` containing
+the respective tests.
+
+Running tests (requires the `pytest` package to be installed):
+
+    pytest
