+++ python/tests/pysb_test_models/bngwiki_egfr_simple_deletemolecules.py
