+++ scripts/run-codecov.sh
@@ -1,18 +1,22 @@
 #!/bin/bash
 # Check code coverage via codecov
 
-SCRIPT_PATH=$(dirname $BASH_SOURCE)
-AMICI_PATH=$(cd $SCRIPT_PATH/.. && pwd)
+script_path=$(dirname $BASH_SOURCE)
+amici_path=$(cd "$script_path"/.. && pwd)
 
-source ${AMICI_PATH}/build/venv/bin/activate
-pip install coverage
+source "${amici_path}"/build/venv/bin/activate
+pip install coverage pytest pytest-cov
 
 if [[ -z "${BNGPATH}" ]]; then
-    export BNGPATH=${AMICI_PATH}/ThirdParty/BioNetGen-2.3.2
+    export BNGPATH="${amici_path}"/ThirdParty/BioNetGen-2.3.2
 fi
 
-python ./tests/testCoverage.py
+pytest \
+  --ignore-glob=*petab* \
+  --cov=amici \
+  --cov-report=xml:"${amici_path}"/coverage_py.xml \
+  --cov-append \
+  "${amici_path}"/python/tests
+
 ret=$?
 if [[ $ret != 0 ]]; then exit $ret; fi
-
-rm -rf ./test_model_steadystate_scaled
