+++ scripts/run-python-tests.sh
@@ -1,23 +1,18 @@
 #!/bin/bash
 # Test python model wrapping inside virtual environment
 
-SCRIPT_PATH=$(dirname $BASH_SOURCE)
-AMICI_PATH=$(cd $SCRIPT_PATH/.. && pwd)
+script_path=$(dirname $BASH_SOURCE)
+amici_path=$(cd "$script_path"/.. && pwd)
 
 set -e
 
 if [[ -z "${BNGPATH}" ]]; then
-    export BNGPATH=${AMICI_PATH}/ThirdParty/BioNetGen-2.3.2
+    export BNGPATH=${amici_path}/ThirdParty/BioNetGen-2.3.2
 fi
 
-cd ${AMICI_PATH}/tests
-source ${AMICI_PATH}/build/venv/bin/activate
-pip install scipy h5py
-python testModels.py
-python testSBML.py
-python testPYSB.py
-python testCPP.py
-python testPreequilibration.py
-python testMisc.py
-python testPandas.py
-python testParameterMapping.py
+cd "${amici_path}"/python/tests
+source "${amici_path}"/build/venv/bin/activate
+pip install scipy h5py pytest pytest-cov
+
+# PEtab tests are run separately
+pytest --ignore-glob=*petab*
