+++ tests/README.md
@@ -1,0 +1,9 @@
+# Directory `tests/`
+
+This directory contains:
+
+- C++ unit tests and integration tests (`cpputest/`)
+- Scripts for running the SBML semantic test suite, exercising the Python
+  interface
+- Scripts for running the PEtab test suite, exercising the Python interface
+- Scripts for testing AMICI performance for importing a somewhat larger model
